Port to Qt4 (using Qt3Support) and Qwt5.
The whole code base is ported to Qt4. The Qt4 port is simplified by using the
'qt3to4' tool and the Qt3Support library provided by Qt4.

The graph plotting code is ported to Qwt5 from Qwt4. Certain functionality is
reduced or removed for now (see note below).

Note that the changes contained within this patch simply do enough to compile
and start up the steerer. It is not necessarily in a state that will be able to
actually steer anything.