Compile the RealityGrid logo into the executable so it can always be loaded.
logo.cpp contains a byte-encoded jpeg file that is loaded and used as if it was
a normal file on disk. The advantage gained by compiling the logo into the
steerer executable is that it doesn't need to be installed where the steerer
can find it each time it is run.

The logo has also been updated to a higher resolution and the old bmp version of
the logo has been removed from the repository.