Update the event model to qt4.
Qt4 requires that QEvents are passed, not QCustomEvents. Can still use
QCustomEvent as it is a subclass of QEvent, but the customEvent method must use
QEvent in its signature rather than QCustomEvent otherwise it doesn't catch the
events that are posted.

This code should probably be rewritten to just use QEvents but, for now, this is
enough to get it working.