Replace Qt3 style dynamic tool-tip with Qt4 event-driven system for the table views.
A "dynamic-tip" class (derived from QToolTip) was used to check the position of
the mouse when over a table and alter the tool-tip text according to which
element it was above. This is not how things should be done in Qt4 so this patch
keeps the code that generates the tool-tip text but overrides the event() method
to capture the mouse postion and create the tool-tip where required.